+++ Changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Protection against synthesis OOC runs in error sending `ipbb` in an infinite loop.
+
 ## [0.4.1] - 2018-02-05
 ### Fixes
 - Minor remaining issues from the splitting of cli commands in interface/implementation addressed.