+++ Changelog.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.4.2] - 2018-02-13
+### Fixes
+- `vivado impl` timing failures severity is now raised to ERROR. 
+
 ### Added
 - Protection against synthesis OOC runs in error sending `ipbb` in an infinite loop.
 - `zsh` autocompletion support.